issue-53873: handling internally any failure related to our own events processing logic (https://issues.jenkins-ci.org/browse/JENKINS-53873)
