+++ src/test/java/com/microfocus/application/automation/tools/octane/PluginAbstractTest.java
@@ -1,5 +1,4 @@
 /*
- *
  *  Certain versions of software and/or documents (“Material”) accessible here may contain branding from
  *  Hewlett-Packard Company (now HP Inc.) and Hewlett Packard Enterprise Company.  As of September 1, 2017,
  *  the Material is now offered by Micro Focus, a separately owned and operated company.  Any reference to the HP
@@ -17,23 +16,36 @@
  * or editorial errors or omissions contained herein.
  * The information contained herein is subject to change without notice.
  * ___________________________________________________________________
- *
  */
 
-package com.microfocus.application.automation.tools.octane.identity;
+package com.microfocus.application.automation.tools.octane;
 
-import com.microfocus.application.automation.tools.octane.PluginAbstractTest;
-import com.microfocus.application.automation.tools.octane.configuration.ConfigurationService;
-import org.junit.Assert;
-import org.junit.Test;
+import com.gargoylesoftware.htmlunit.html.HtmlForm;
+import com.gargoylesoftware.htmlunit.html.HtmlPage;
+import com.hp.octane.integrations.dto.DTOFactory;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.jvnet.hudson.test.JenkinsRule;
 
-public class ServerIdentityTest extends PluginAbstractTest {
+import java.util.UUID;
 
-	@Test
-	public void testIdentity() {
-		ConfigurationService.getAllSettings().forEach(settings -> {
-			Assert.assertNotNull(settings.getIdentity());
-			Assert.assertFalse(settings.getIdentity().isEmpty());
-		});
+public class PluginAbstractTest {
+	protected static final DTOFactory dtoFactory = DTOFactory.getInstance();
+	protected static String ssp;
+
+	@ClassRule
+	public static final JenkinsRule rule = new JenkinsRule();
+	public static final JenkinsRule.WebClient client = rule.createWebClient();
+
+
+	@BeforeClass
+	public static void init() throws Exception {
+		HtmlPage configPage = client.goTo("configure");
+		HtmlForm form = configPage.getFormByName("config");
+		ssp = UUID.randomUUID().toString();
+		form.getInputByName("_.uiLocation").setValueAttribute("http://localhost:8008/ui/?p=" + ssp + "/1002");
+		form.getInputByName("_.username").setValueAttribute("username");
+		form.getInputByName("_.password").setValueAttribute("password");
+		rule.submit(form);
 	}
 }
