+++ src/test/java/com/microfocus/application/automation/tools/octane/tests/gherkin/GherkinResultsTest.java
@@ -22,7 +22,7 @@
 
 package com.microfocus.application.automation.tools.octane.tests.gherkin;
 
-import com.microfocus.application.automation.tools.octane.PlugInAbstractTest;
+import com.microfocus.application.automation.tools.octane.PluginAbstractTest;
 import com.microfocus.application.automation.tools.octane.actions.cucumber.CucumberTestResultsActionPublisher;
 import com.microfocus.application.automation.tools.octane.tests.CopyResourceSCM;
 import com.microfocus.application.automation.tools.octane.tests.TestUtils;
@@ -55,7 +55,7 @@ import java.util.UUID;
  * Created by franksha on 05/01/2017.
  */
 @SuppressWarnings({"squid:S2699","squid:S3658","squid:S2259","squid:S1872","squid:S2925","squid:S109","squid:S1607","squid:S2701","squid:S2698"})
-public class GherkinResultsTest extends PlugInAbstractTest {
+public class GherkinResultsTest extends PluginAbstractTest {
 
     private static String mavenName;
